fix(storybook): updated storybook to version 10
