+++ package.json
@@ -39,7 +39,7 @@
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/user-event": "^14.5.2",
-    "@vitest/ui": "^2.1.4",
+    "@vitest/ui": "^4.1.8",
     "cli-progress": "^3.12.0",
     "concurrently": "^9.1.2",
     "core-js": "^3.39.0",
@@ -78,7 +78,7 @@
     "typescript-eslint": "^8.13.0",
     "vinyl-buffer": "^1.0.1",
     "vinyl-source-stream": "^2.0.0",
-    "vitest": "^2.1.4",
+    "vitest": "^4.1.8",
     "wait-on": "^8.0.2",
     "whatwg-fetch": "^3.6.20"
   },
