+++ packages/react/package.json
@@ -57,9 +57,9 @@
   },
   "devDependencies": {
     "@types/react": "^18.3.18",
-    "@vitejs/plugin-react": "^4.3.4",
+    "@vitejs/plugin-react": "^6.0.2",
     "typescript": "^5.7.3",
-    "vite": "^6.0.11",
+    "vite": "^8.0.16",
     "vite-plugin-dts": "^4.5.0"
   },
   "peerDependencies": {
