--- packages/storybook-email/.storybook/main.js
@@ -1,35 +1,0 @@
-import { dirname, join } from "path";
-
-/** @type { import('@storybook/html-webpack5').StorybookConfig } */
-const config = {
-  stories: [
-    "../stories/**/*.mdx",
-    "../stories/**/*.@(stories|story).@(js|jsx|ts|tsx)",
-  ],
-
-  addons: [
-    getAbsolutePath("@storybook/addon-a11y"),
-    getAbsolutePath("@storybook/addon-links"),
-    getAbsolutePath("@storybook/addon-docs"),
-  ],
-
-  features: {
-    postcss: false,
-    previewMdx2: false,
-  },
-
-  framework: {
-    name: getAbsolutePath("@storybook/html-vite"),
-    options: {},
-  },
-};
-
-/**
- * This function is used to resolve the absolute path of a package.
- * It is needed in projects that use Yarn PnP or are set up within a monorepo.
- */
-function getAbsolutePath(value) {
-  return dirname(require.resolve(join(value, "package.json")));
-}
-
-export default config;
