+++ packages/storybook-email/.storybook/main.mjs
@@ -1,10 +1,14 @@
-import { join, dirname } from "path";
+import { dirname, join } from "path";
+import { fileURLToPath } from "url";
 
-/** @type { import('@storybook/react-vite').StorybookConfig } */
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+/** @type { import('@storybook/html-vite').StorybookConfig } */
 const config = {
   stories: [
     "../stories/**/*.mdx",
-    "../stories/**/*.stories.@(js|jsx|mjs|ts|tsx)",
+    "../stories/**/*.@(stories|story).@(js|jsx|ts|tsx)",
   ],
 
   addons: [
@@ -13,23 +17,18 @@ const config = {
     getAbsolutePath("@storybook/addon-docs"),
   ],
 
-  staticDirs: ["../../storybook-html/public"],
-
   framework: {
-    name: getAbsolutePath("@storybook/react-vite"),
+    name: getAbsolutePath("@storybook/html-vite"),
     options: {},
   },
-
-  typescript: {
-    reactDocgen: "react-docgen-typescript",
-  },
 };
-export default config;
 
 /**
  * This function is used to resolve the absolute path of a package.
  * It is needed in projects that use Yarn PnP or are set up within a monorepo.
  */
 function getAbsolutePath(value) {
-  return dirname(require.resolve(join(value, "package.json")));
+  return dirname(fileURLToPath(import.meta.resolve(value)));
 }
+
+export default config;
