+++ packages/storybook-email/package.json
@@ -23,27 +23,28 @@
     "@babel/preset-react": "^7.25.9",
     "@babel/preset-typescript": "^7.26.0",
     "@material/circular-progress": "^14.0.0",
-    "@storybook/addon-a11y": "9.1.20",
-    "@storybook/addon-docs": "9.1.20",
-    "@storybook/addon-links": "9.1.20",
-    "@storybook/html-vite": "9.1.20",
+    "@storybook/addon-a11y": "^10.4.3",
+    "@storybook/addon-docs": "^10.4.3",
+    "@storybook/addon-links": "^10.4.3",
+    "@storybook/html-vite": "^10.4.3",
     "@uqds/accordion": "^7.0.5",
     "babel-loader": "^9.2.1",
     "core-js": "^3.39.0",
     "css-loader": "^7.1.2",
+    "eslint-plugin-storybook": "^10.4.3",
     "font-awesome": "^4.7.0",
     "http-server": "^14.1.1",
     "joi-browser": "^13.4.0",
     "mermaid": "^11.4.0",
     "sass": "^1.89.2",
     "sass-loader": "^16.0.3",
-    "storybook": "9.1.20",
+    "storybook": "^10.4.3",
     "tocbot": "^4.31.0",
     "ts-loader": "^9.5.1",
     "typescript": "^5.6.3",
     "underscore": "^1.13.7",
-    "zxcvbn": "^4.4.2",
-    "eslint-plugin-storybook": "9.1.20"
+    "vite": "^8.0.16",
+    "zxcvbn": "^4.4.2"
   },
   "dependencies": {
     "classnames": "^2.5.1"
