+++ packages/storybook-react/package.json
@@ -21,13 +21,13 @@
     "@uqds/react": "^2.0.7"
   },
   "devDependencies": {
-    "@storybook/addon-a11y": "9.1.20",
-    "@storybook/addon-docs": "9.1.20",
-    "@storybook/addon-links": "9.1.20",
-    "@storybook/react-vite": "9.1.20",
-    "@vitejs/plugin-react": "^4.3.4",
-    "storybook": "9.1.20",
-    "vite": "^6.0.11",
-    "eslint-plugin-storybook": "9.1.20"
+    "@storybook/addon-a11y": "^10.4.3",
+    "@storybook/addon-docs": "^10.4.3",
+    "@storybook/addon-links": "^10.4.3",
+    "@storybook/react-vite": "^10.4.3",
+    "@vitejs/plugin-react": "^6.0.2",
+    "eslint-plugin-storybook": "^10.4.3",
+    "storybook": "^10.4.3",
+    "vite": "^8.0.16"
   }
 }
